Refactor: [#46, #35] EventCard.js 컴포넌트 생성, EventPage Presenter/Container 컴포넌트 구조 제거
